Added basic support for reading the workout HTML elements
This solution contains CLI and Core projects. Core contains
the actual Crawler which opens a Firefox application and
automates the fetching of workouts. Currenty this supports
only OSX since the Firefox location is hardcoded.

The CLI project is just a console frontend which uses
Mono.Options to get the username and password of the person's
fitocracy account. After validation and verification it
passes it over to the Crawler